Move capabilities.ts under vm which it's more closely tied to
