+++ plugin-server/src/worker/vm/capabilities.ts
