+++ plugin-server/tests/worker/vm/capabilities.test.ts
